refactor : 최종적인 코드 구조 변경(package 변경)
