+++ src/main/java/bridge/Direction.java
@@ -1,4 +1,4 @@
-package utils;
+package bridge;
 
 public enum Direction {
