+++ src/main/java/game/Controller.java
@@ -1,7 +1,6 @@
 package game;
 
 import inMemoryDB.GameData;
-import utils.GameResult;
 import view.InputView;
 import view.OutputView;
 import java.util.List;
