+++ src/main/java/game/GameResult.java
@@ -1,4 +1,4 @@
-package utils;
+package game;
 
 public enum GameResult {
