+++ src/main/java/game/Restart.java
@@ -1,4 +1,4 @@
-package utils;
+package game;
 
 public enum Restart {
